+++ hips/utils/__init__.py
@@ -5,5 +5,3 @@ Note that those are mostly for internal notes, if you use them from
 external scripts or packages at all, you should not expect stability
 from our side and it might be better if you copy over what you need.
 """
-from .wcs import *
-from .healpix import *